pyproject.toml
[project]
name = "fs-builder"
version = "0.3.0"
description = "CLI-first demo for natural language to Onshape FeatureScript"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.23.0",
    "openai>=1.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.0",
]

[project.scripts]
fs-builder = "fs_builder.cli:main"

[build-system]
requires = ["setuptools>=58", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"fs_builder.prompts" = ["*.txt"]
"fs_builder.references" = ["*.md"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"

[tool.ruff]
target-version = "py39"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
mypy_path = ["src"]
files = ["src"]
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_ignores = true
ignore_missing_imports = true
